# legacy-nextjs-app/openapi.yaml
openapi: 3.0.3
info:
  title: KlipCam API
  description: API for KlipCam, a Creator AI platform.
  version: 1.0.0
servers:
  - url: /api
security:
  - ClerkAuth: []
paths:
  /jobs:
    post:
      summary: Create a new generation job
      description: Creates a new image or video generation job after validating user credits.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [image, video]
                subtype:
                  type: string
                  enum: [t2i, i2i, t2v, i2v, spider, lora_train, lora_t2i, lora_i2i]
                params:
                  type: object
                  description: Model-specific parameters.
                  properties:
                    prompt:
                      type: string
                    negative_prompt:
                      type: string
                    width:
                      type: integer
                    height:
                      type: integer
                    steps:
                      type: integer
                    guidance:
                      type: number
                    seed:
                      type: integer
                    reference_image_url:
                      type: string
                      format: uri
                    model_tier:
                      type: string
                      enum: [base, premium]
                    aspect:
                      type: string
                      enum: ["9:16", "1:1", "portrait"]
                    duration:
                      type: integer
                      default: 3
                    lora_model_id:
                      type: string
                      format: uuid
                      description: LoRA model ID for enhanced generation
                    lora_trigger_word:
                      type: string
                      description: Trigger word for LoRA model activation
              required:
                - type
                - subtype
                - params
      responses:
        '202':
          description: Job accepted and queued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          description: Bad request (e.g., invalid parameters).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '402':
          description: Insufficient credits.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: Get user's job history
      description: Returns a paginated list of the user's generation jobs.
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: status
          in: query
          schema:
            type: string
            enum: [queued, running, succeeded, failed]
          description: Filter jobs by status
        - name: type
          in: query
          schema:
            type: string
            enum: [image, video]
          description: Filter jobs by type
      responses:
        '200':
          description: A paginated list of jobs.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Job'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /replicate/webhook:
    post:
      summary: Replicate webhook
      description: Handles completion callbacks from the Replicate API. This endpoint is not called by the frontend client.
      requestBody:
        description: Replicate prediction object.
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook received and processed.
        '400':
          description: Bad request or invalid signature.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /jobs/{id}:
    get:
      summary: Get specific job details
      description: Returns detailed information about a specific job.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Job not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Job:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [image, video]
        subtype:
          type: string
          enum: [t2i, i2i, t2v, i2v, spider, upscale, lora_train, lora_t2i, lora_i2i]
        model:
          type: string
          description: The Replicate model used for generation
        params:
          type: object
          description: Generation parameters used for the job
        status:
          type: string
          enum: [queued, running, succeeded, failed]
        cost_credits:
          type: integer
          minimum: 1
          description: Credits charged for this job
        replicate_prediction_id:
          type: string
          nullable: true
          description: Replicate prediction ID for tracking
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        error:
          type: string
          nullable: true
          description: Error message if job failed
      required:
        - id
        - user_id
        - type
        - subtype
        - status
        - cost_credits
        - created_at
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error type or code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error context
      required:
        - error
        - message
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          description: Array of items
        pagination:
          type: object
          properties:
            total:
              type: integer
              description: Total number of items
            limit:
              type: integer
              description: Items per page
            offset:
              type: integer
              description: Current offset
            has_more:
              type: boolean
              description: Whether there are more items
          required:
            - total
            - limit
            - offset
            - has_more
      required:
        - data
        - pagination
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk JWT token. Include in Authorization header as "Bearer <token>"